# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zanichelli API IdentityProvider",
        "description": "REST APIs to use SSO (single-sign-on) for applications ",
        "contact": {
            "name": "Zanichelli DEV team",
            "email": "gsiciliano@zanichelli.it"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "list of roles",
                "description": "Returns the entire list of roles",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "RoleController.create",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Role name",
                                        "type": "string",
                                        "example": "ADMIN_IDP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Invalid scope or client role, Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "manage-idp"
                        ]
                    }
                ]
            }
        },
        "/v1/roles/{id}": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Remove role by id",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "RoleController.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Error on deleting",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Invalid scope or client role, Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "manage-idp"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "http://localhost:8083/oauth/token",
                        "scopes": []
                    }
                },
                "scheme": [
                    "http",
                    "https"
                ]
            }
        }
    },
...
}